% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unvec_symm}
\alias{unvec_symm}
\title{Inverse of \code{vec_symm}}
\usage{
unvec_symm(x, dim)
}
\arguments{
\item{x}{A numeric vector of length \code{n*(n+1)/2}.}

\item{dim}{The matrix dimension \code{n}.}
}
\value{
The corresponding \verb{n x n} symmetric matrix.
}
\description{
Inverse of \code{vec_symm}
}
