% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffcp-package.R
\docType{package}
\name{diffcp-package}
\alias{diffcp-package}
\alias{diffcp}
\title{diffcp: Differentiating Through Cone Programs}
\description{
R port of the Python \code{diffcp} package (Agrawal, Barratt, Boyd 2019).
Computes the derivative of the optimal solution map of a convex cone
program with respect to its data, via the implicit-function theorem
on the homogeneous self-dual embedding.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bnaras/diffcp}
  \item Report bugs at \url{https://github.com/bnaras/diffcp/issues}
}

}
\author{
\strong{Maintainer}: Balasubramanian Narasimhan \email{naras@stanford.edu}

Authors:
\itemize{
  \item Akshay Agrawal
  \item Shane Barratt
  \item Stephen Boyd
  \item Enzo Busseti
  \item Walaa Moursi
}

}
\keyword{internal}
